New function wrappers (#21)
* Simplify python usage

* Fix layer query

* Fix warnings

* Ignore test files

* Update C++ test case

* Query as class

* New function wrapper

* Update Python test case

+++ include/function.h
@@ -12,11 +12,11 @@ public:
     std::string name;
     std::vector<Event *> cuda_ptr_list;
     float fp32active_sum;
-    float duration_sum;
+    int duration_sum;
     Function(const std::string &name,
              const std::vector<Event *> &cuda_ptr_list,
              const float &fp32active_sum,
-             const float &duration_sum);
+             const int &duration_sum);
     ~Function();
 };
